feat(web) - ? key opens help modal
